Control large amplitude / squeal on higher notes
Larger volume is due to a combination of the initial bellFreq combs, and the later delaySpike combs interacting. Dial down the earlier combs with higher frequencies (the .explin), and avoid high frequencies for the delaySpike combs. The latter is too messy at the moment, but it's better than the squealing.