Refactor hiss component of \driver
This makes the hiss build-up more snappy, and injects some of the hiss sound driven by the core driver pulse when noiseAmt is turned up. The two (hissPulse and hissBuild) don't sound well together, but sound good on their own, so at some point it will have to be one or the other.